B-1 adding exception handler to controllers.
