+++ auth/src/main/java/net/prottonne/lab/auth/constant/EnviromentConstant.java
@@ -1,19 +1,18 @@
 /*
  * Copyright 2017 Prottonne
  */
-package net.prottonne.lab.auth.controller;
+package net.prottonne.lab.auth.constant;
 
 /**
  *
  * @author Jose Osuna
  */
-public enum ErrorMessage {
-    CANNOT_BE_PROCESSED("cannot be processed"),
-    ACCESS_DENIED("access denied");
+public enum EnviromentConstant {
+    JWT_SECRET("WWW_1232");
 
     private final String value;
 
-    private ErrorMessage(String value) {
+    private EnviromentConstant(String value) {
         this.value = value;
     }
